pyproject.toml
[tool.poetry]
name = "gozle-ads"
version = "0.1.0"
description = ""
authors = ["Rowsen <bayramovw0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
django-debug-toolbar = "^4.2.0"
psycopg2 = "^2.9.7"
pillow = "^10.0.1"
channels = "^4.0.0"
daphne = "^4.0.0"
drf-extra-fields = "^3.7.0"
drf-spectacular = "^0.26.5"
django-multiselectfield = "^0.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
